+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * Kubernetes support
 * [EOS-4775] Fix step to remove user/group from tenant
 * [SPK-1168] Get log from mesos if CCT doesn't return log path
+* [DGPB-6794] Add support to Governance service role
 
 ## 0.20.0 (December 17, 2020)